backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.1
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.12
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.3
whitenoise==6.9.0
